Update 2
Added a new feature to the program - searching for changes in files,
creating or deleting a file in the directory to be scanned. Now, if the
result file has already been created, you can rescan the directory and
find the changes (and then output to console);
optimized Secure Hash Algorithm 1 (now hashing file path and date of
last change, not file path only).